requirements.txt
aiogram==3.2.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
cachetools==5.3.2
certifi==2023.11.17
colorama==0.4.4
cryptography==3.4.8
frozenlist==1.4.1
importlib-metadata==4.6.4
launchpadlib==1.10.16
magic-filter==1.0.12
monotonic==1.6
multidict==6.0.4
oauthlib==3.2.0
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
protobuf==3.12.4
psycopg2==2.9.9
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
pytweening==1.0.7
pytz==2022.1
redis==5.0.1
reportlab==3.6.8
requests==2.25.1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==1.26.5
wadllib==1.3.6
python-dotenv==1.0.0
asyncpg==0.29.0
